feat: add gas used and transactions fee
